requirements.txt
Flask==3.0.2
Werkzeug>=3.0.0
Pillow>=10.4.0
numpy>=1.26.0
opencv-python-headless>=4.8.0.74
PyMuPDF==1.24.9
